% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinatePlane.R
\name{CoordinatePlane}
\alias{CoordinatePlane}
\title{Plots an empty coordinate (cartesian) plane with customizable limits for the X and Y axis}
\usage{
CoordinatePlane(x_min, x_max, y_min, y_max)
}
\arguments{
\item{x_min}{Lowest value for the X axis}

\item{x_max}{Highest value for the X axis}

\item{y_min}{Lowest value for the Y axis}

\item{y_max}{Highest value for the Y axis}
}
\value{
None. It produces a plot of a coordinate plane with axes and grid
}
\description{
\code{CoordinatePlane} plots an empty coordinate (cartesian) plane with customizable limits for the X and Y axis.
}
\examples{
x_min <- -5
x_max <- 5
y_min <- -5
y_max <- 5
CoordinatePlane(x_min, x_max, y_min, y_max)
}
