% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLinePoints.R
\name{CreateLinePoints}
\alias{CreateLinePoints}
\title{Creates a vector that represents the line that connects two points}
\usage{
CreateLinePoints(P1, P2)
}
\arguments{
\item{P1}{Vector containing the xy-coordinates of point 1}

\item{P2}{Vector containing the xy-coordinates of point 2}
}
\value{
Returns a vector which contains the slope and intercept of the defined line. If the points have the same x-coordinate, the slope is set to \code{Inf} and the intercept is replaced by the x-value for the line (which is a vertical line in this situation)
}
\description{
\code{CreateLinePoints} creates a vector that represents the line that connects two points
}
\examples{
P1 <- c(0,0)
P2 <- c(1,1)
Line <- CreateLinePoints(P1, P2)
}
