% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePolygon.R
\name{CreatePolygon}
\alias{CreatePolygon}
\title{Creates a matrix to represent the polygon that connects several points}
\usage{
CreatePolygon(...)
}
\arguments{
\item{...}{An undetermined number of points introduced by the user in the form of vectors}
}
\value{
Returns a matrix which contains the points of the polygon. Each row represents one of the points
}
\description{
\code{CreatePolygon} creates a matrix to represent the polygon that connects several points
}
\examples{
P1 <- c(0,0)
P2 <- c(1,1)
P3 <- c(2,0) 
Poly <- CreatePolygon(P1, P2, P3)
}
