% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRegularPolygon.R
\name{CreateRegularPolygon}
\alias{CreateRegularPolygon}
\title{Creates a matrix to represent a regular polygon}
\usage{
CreateRegularPolygon(n, C, l)
}
\arguments{
\item{n}{Number of sides for the polygon}

\item{C}{Vector containing the xy-coordinates for the center of the regular polygon}

\item{l}{Length of the sides for the polygon}
}
\value{
Returns a matrix which contains the points of a regular polygon given its number of points and the length of its sides. Each row represents one of the points
}
\description{
\code{CreateRegularPolygon} creates a matrix to represent the polygon that connects several points
}
\examples{
x_min <- -5
x_max <- 5
y_min <- -5
y_max <- 5
CoordinatePlane(x_min, x_max, y_min, y_max)
n <- 5
C <- c(0,0) 
l <- 1
Penta <- CreateRegularPolygon(n, C, l)
Draw(Penta, "blue", label = TRUE)
}
