% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSegmentAngle.R
\name{CreateSegmentAngle}
\alias{CreateSegmentAngle}
\title{Creates a matrix that represents the segment that starts from a point with certain length and angle}
\usage{
CreateSegmentAngle(P, angle, l)
}
\arguments{
\item{P}{Vector containing the xy-coordinates of the point}

\item{angle}{Angle in degrees (0-360) for the segment}

\item{l}{Positive number that indicates the length for the segment}
}
\value{
Returns a matrix which contains the points that determine the extremes of the segment
}
\description{
\code{DrawSegment} plots the segment that connects two points in a previously generated coordinate plane
}
\examples{
x_min <- -5
x_max <- 5
y_min <- -5
y_max <- 5
CoordinatePlane(x_min, x_max, y_min, y_max)
P <- c(0,0)
angle <- 30
l <- 1
Segment <- CreateSegmentAngle(P, angle, l)
Draw(Segment, "black")
}
