% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSegmentPoints.R
\name{CreateSegmentPoints}
\alias{CreateSegmentPoints}
\title{Creates a matrix that represents the segment that connects two points}
\usage{
CreateSegmentPoints(P1, P2)
}
\arguments{
\item{P1}{Vector containing the xy-coordinates of point 1}

\item{P2}{Vector containing the xy-coordinates of point 2}
}
\value{
Returns a matrix which contains the points that determine the extremes of the segment
}
\description{
\code{DrawSegment} plots the segment that connects two points in a previously generated coordinate plane
}
\examples{
x_min <- -5
x_max <- 5
y_min <- -5
y_max <- 5
CoordinatePlane(x_min, x_max, y_min, y_max)
P1 <- c(0,0)
P2 <- c(1,1)
Segment <- CreateSegmentPoints(P1, P2)
Draw(Segment, "black")
}
