% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceLines.R
\name{DistanceLines}
\alias{DistanceLines}
\title{Computes the distance between two lines}
\usage{
DistanceLines(Line1, Line2)
}
\arguments{
\item{Line1}{Line object previously created with \code{CreateLinePoints} or \code{CreateLineAngle}}

\item{Line2}{Line object previously created with \code{CreateLinePoints} or \code{CreateLineAngle}}
}
\value{
Returns the distance between two points
}
\description{
\code{DistanceLines} computes the distance between two lines
}
\examples{
P1 <- c(0,0)
P2 <- c(1,1)
Line1 <- CreateLinePoints(P1, P2)
P3 <- c(1,-1)
P4 <- c(2,0)
Line2 <- CreateLinePoints(P3, P4)
d <- DistanceLines(Line1, Line2)
}
