% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinesAngles.R
\name{LinesAngles}
\alias{LinesAngles}
\title{Computes the angle that form two lines}
\usage{
LinesAngles(Line1, Line2)
}
\arguments{
\item{Line1}{Line object previously created with \code{CreateLinePoints} or \code{CreateLineAngle}}

\item{Line2}{Line object previously created with \code{CreateLinePoints} or \code{CreateLineAngle}}
}
\value{
Returns the angle that form the two lines
}
\description{
\code{LinesAngles} computes the angle that form two lines
}
\examples{
P1 <- c(0,0)
P2 <- c(1,1)
Line1 <- CreateLinePoints(P1, P2)
P3 <- c(1,-1)
P4 <- c(2,3)
Line2 <- CreateLinePoints(P3, P4)
angle <- LinesAngles(Line1, Line2)
}
