% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MidPoint.R
\name{MidPoint}
\alias{MidPoint}
\title{Computes the middle point of the segment that connects two points}
\usage{
MidPoint(P1, P2)
}
\arguments{
\item{P1}{Vector containing the xy-coordinates of point 1}

\item{P2}{Vector containing the xy-coordinates of point 2}
}
\value{
Returns a vector containing the xy-coordinates of the middle point of the segment that connects \code{P1} and \code{P2}
}
\description{
\code{MidPoint} computes the middle point of the segment that connects two points
}
\examples{
P1 <- c(0,0)
P2 <- c(1,1)
mid <- MidPoint(P1, P2)
}
