% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolygonAngles.R
\name{PolygonAngles}
\alias{PolygonAngles}
\title{Computes each of the existing angles in a given polygon}
\usage{
PolygonAngles(Poly)
}
\arguments{
\item{Poly}{Polygon object, previously created with function \code{CreatePolygon}}
}
\value{
Returns a vector containing the angles for each of the points of a polygon. The resulting vector follows the order of the points in the defined polygon
}
\description{
\code{PolygonAngles} computes each of the existing angles in a given polygon
}
\examples{
P1 <- c(0,0)
P2 <- c(1,1)
P3 <- c(2,0) 
Poly <- CreatePolygon(P1, P2, P3)
angles <- PolygonAngles(Poly)
}
