% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectPoints.R
\name{SelectPoints}
\alias{SelectPoints}
\title{Selection of points from the coordinate plane}
\usage{
SelectPoints(n)
}
\arguments{
\item{n}{Number of points to select from the current coordinate plane}
}
\value{
Returns a vector or matrix which contains the xy-coordinates of the selected points. Each row represents one of the points. If \code{n = 1} the output is a numeric vector, if \code{n = 2} then it is a \code{Segment}, and for \code{n > 2} the object is a polygon.
}
\description{
\code{SelectPoints} allows the selection of points from the coordinate plane
}
\examples{
n <- 3
points <- SelectPoints(n)
}
